GoGoDuck refactoring
 * URLMangler class
 * Instantiate classes with GoGoDuckModule.newInstance()
 * Thread safety in SubsetParameters class
